# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TPayments API Documentation",
        "description": "Swagger doc for tpayments",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://turkmenpost.loc/api/v1/",
            "description": "TPayments API Server"
        }
    ],
    "paths": {
...
        "/payments/gts": {
            "get": {
                "tags": [
                    "GTS"
                ],
                "summary": "Get balance of gts user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/GTSGetBalanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "406": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "status",
                                            "description": "result status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "title": "balance",
                                                    "description": "balance of gts user",
                                                    "type": "integer",
                                                    "example": "11.24"
                                                },
                                                "agreement_number": {
                                                    "title": "agreement_number",
                                                    "description": "agreement_number of this user, you should use this number in top up balance request",
                                                    "type": "integer",
                                                    "example": "inet-332233"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "GTS"
                ],
                "summary": "Create a new GTS payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/GTSPaymentCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "406": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "status",
                                            "description": "result status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "title": "data",
                                            "description": "Created payment data",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaymentModel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payments/telecom": {
            "get": {
                "tags": [
                    "TELECOM"
                ],
                "summary": "Get payable status of telecom user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/TelecomCheckRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "406": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "status",
                                            "description": "request result status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "result": {
                                                    "title": "result",
                                                    "description": "true if phone number can be payed",
                                                    "type": "boolean",
                                                    "example": "false"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TELECOM"
                ],
                "summary": "Create a new telecom payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/TelecomPaymentCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "406": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "status",
                                            "description": "result status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "title": "data",
                                            "description": "Created payment data",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaymentModel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentMethod": {
                "title": "PaymentMethod",
                "description": "Used method for payment",
                "type": "string",
                "enum": [
                    "CARD"
                ],
                "example": "CARD"
            },
            "PaymentState": {
                "title": "PaymentState",
                "description": "Main state of payment",
                "type": "string",
                "enum": [
                    "CREATED",
                    "PENDING",
                    "SUCCESS",
                    "FAILED"
                ],
                "example": "ORDERED"
            },
            "PaymentModel": {
                "title": "PaymentModel",
                "description": "Payment model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "The internal ID of this payment",
                        "type": "integer",
                        "format": "int64",
                        "example": "1443"
                    },
                    "uuid": {
                        "title": "uuid",
                        "description": "The internal UUID of this payment",
                        "type": "string",
                        "format": "uuid",
                        "example": "94387003-b27d-4e20-b157-b35079230580"
                    },
                    "ref_uuid": {
                        "title": "ref_uuid",
                        "description": "reference uuid provided by client",
                        "type": "string",
                        "format": "uuid",
                        "example": "94387003-b27d-4e20-b157-b35079230580"
                    },
                    "client_id": {
                        "title": "client_id",
                        "description": "Creator identifier of the payment ",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "state": {
                        "$ref": "#/components/schemas/PaymentState"
                    },
                    "method": {
                        "$ref": "#/components/schemas/PaymentMethod"
                    },
                    "service": {
                        "$ref": "#/components/schemas/ServicesEnum"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ServicesTypeEnum"
                    },
                    "amount": {
                        "title": "Amount",
                        "description": "The amount of payment in cents",
                        "type": "integer",
                        "format": "int64",
                        "example": "12000"
                    },
                    "created_at": {
                        "title": "RFC3339",
                        "description": "Created datetime of payment",
                        "type": "date-time",
                        "example": "2020-05-11T07:35:49.00Z"
                    },
                    "updated_at": {
                        "title": "RFC3339",
                        "description": "Updated datetime of payment",
                        "type": "date-time",
                        "example": "2020-05-11T07:35:49.00Z"
                    },
                    "completed_at": {
                        "title": "RFC3339",
                        "description": "Payment completed date-time",
                        "type": "date-time",
                        "example": "2020-05-11T07:35:49.00Z"
                    }
                },
                "type": "object"
            },
            "ServicesEnum": {
                "title": "ServicesEnum",
                "description": "Service types",
                "type": "string",
                "enum": [
                    "GTS",
                    "TELECOM",
                    "TMCELL"
                ],
                "example": "GTS"
            },
            "ServicesTypeEnum": {
                "title": "ServicesTypeEnum",
                "description": "Service subtypes",
                "type": "string",
                "enum": [
                    "GTS_INET",
                    "GTS_PHONE",
                    "GTS_IPTV",
                    "TELC_INET",
                    "TMCELL_PHONE"
                ],
                "example": "GTS_INET"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}